update operation with otder items update

+++ order-mgt/src/main/java/com/suinfinity/order/service/OrderService.java
@@ -2,6 +2,7 @@ package com.suinfinity.order.service;
 
 import com.suinfinity.order.dto.OrderDTO;
 import com.suinfinity.order.dto.OrderItemDTO;
+import com.suinfinity.order.dto.OrderItemResponseDTO;
 import com.suinfinity.order.dto.OrderResponseDTO;
 import com.suinfinity.order.exception.OrderNotFoundException;
 import com.suinfinity.order.mapper.OrderItemMapper;
@@ -14,6 +15,7 @@ import com.suinfinity.order.util.OrderUtil;
 import java.time.Instant;
 import java.time.ZoneOffset;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.ListIterator;
@@ -37,12 +39,7 @@ public class OrderService {
     List<Map<String, String>> orderItems = orderDTO.getOrderItems();
     try {
       orderRepository.save(order);
-      for (Map<String, String> orderItemMap : orderItems) {
-        OrderItemDTO orderItemDTO = OrderUtil.getOrderItemDTO(orderItemMap);
-        OrderItem orderItem = OrderItemMapper.INSTANCE.toOrderItem(orderItemDTO);
-        orderItem.setOrderId(order.getOrderId());
-        orderItemRepository.save(orderItem);
-      }
+      this.saveOrderItems(orderItems, order);
     } catch (RuntimeException e) {
       log.error("Order {} placement failed ", order.getOrderId());
       throw new RuntimeException("Order " + "'" + order.getOrderId() + "'" + " placement failed");
@@ -71,6 +68,7 @@ public class OrderService {
     List<Order> orders = orderRepository.findAll();
     ListIterator<Order> listIterator = orders.listIterator();
     Map<Long, List<Map<String, String>>> orderItemDTOSMap = new HashMap<>();
+
     while (listIterator.hasNext()) {
       Order order = listIterator.next();
       long orderId = order.getOrderId();
@@ -85,7 +83,7 @@ public class OrderService {
     }
     if (orders.isEmpty()) {
       log.info("Orders are empty");
-      return ResponseEntity.noContent().build();
+      return ResponseEntity.notFound().build();
     }
     return ResponseEntity.ok(orderResponseDTOS);
   }
@@ -100,33 +98,66 @@ public class OrderService {
                   return new OrderNotFoundException(
                       "Update operation failed, Order " + "'" + orderId + "'" + " does not exist");
                 });
-    orderDTO.setOrderDate(Instant.now().atZone(ZoneOffset.UTC).toLocalDateTime());
-    order.setAmount(orderDTO.getAmount());
-    try {
-      orderRepository.save(order);
-    } catch (RuntimeException e) {
-      throw new RuntimeException("Order " + "'" + orderId + "'" + " update failed");
+    Order exsisOrder = new Order();
+    List<OrderItemDTO> orderItemDTOS = new ArrayList<>();
+    for (Map<String, String> orderItemMap : orderDTO.getOrderItems()) {
+      orderItemDTOS.add(OrderUtil.getOrderItemDTO(orderItemMap));
     }
-    OrderResponseDTO orderResponseDTO = OrderMapper.INSTANCE.toOrderResponseDTO(order);
+
+    List<OrderItem> exsistOrderItemList = orderItemRepository.findByOrderId(orderId);
+    for (int i = 0; i < exsistOrderItemList.size(); i++) {
+      OrderItem item = exsistOrderItemList.get(i);
+      OrderItemDTO orderItemDTO = orderItemDTOS.get(i);
+      item.setProductId(orderItemDTO.getProductId());
+      item.setPrice(orderItemDTO.getPrice());
+      item.setQuantity(orderItemDTO.getQuantity());
+      orderItemRepository.save(item);
+    }
+    Date updatedDate =
+        Date.from(Instant.now().atZone(ZoneOffset.UTC).toLocalDateTime().toInstant(ZoneOffset.UTC));
+    order.setOrderDate(updatedDate);
+    order.setAmount(orderDTO.getAmount());
+    orderRepository.save(order);
+    log.info("Order updated successfully");
+    OrderResponseDTO orderResponseDTO = OrderMapper.INSTANCE.toOrderResponseDTO(exsisOrder);
     return ResponseEntity.ok(orderResponseDTO);
   }
 
   @Transactional
   public ResponseEntity<Void> deleteOrder(long orderId) {
-    orderItemRepository.deleteByOrderId(orderId);
-    orderRepository.deleteById(orderId);
+    long orderCount = orderRepository.deleteByOrderId(orderId);
+    if (orderCount == 0) {
+      log.error("Order id '{}' does not exist, delete operation failed", orderId);
+      throw new RuntimeException(
+          "Delete operation failed, order id:" + "'" + orderId + "'" + " does not exist");
+    }
+    long itemCount = orderItemRepository.deleteByOrderId(orderId);
+    log.info("Deleted order item count is: '{}' :", itemCount);
     return ResponseEntity.noContent().build();
   }
 
   private List<Map<String, String>> getOrderItems(List<OrderItem> orderItemList) {
     ListIterator<OrderItem> listIterator = orderItemList.listIterator();
-    List<Map<String, String>> orderItemDTOS = new ArrayList<>();
+    List<Map<String, String>> orderItemResponseDTOList = new ArrayList<>();
     while (listIterator.hasNext()) {
       OrderItem orderItem = listIterator.next();
-      OrderItemDTO orderItemDTO = OrderItemMapper.INSTANCE.toOrderItemDTO(orderItem);
-      Map<String, String> orderItems = OrderUtil.setOrderItemDTO(orderItemDTO);
-      orderItemDTOS.add(orderItems);
+      OrderItemResponseDTO orderItemResponseDTO =
+          OrderItemMapper.INSTANCE.toOrderItemDTO(orderItem);
+      Map<String, String> orderItems = OrderUtil.setOrderItemDTO(orderItemResponseDTO);
+      orderItemResponseDTOList.add(orderItems);
     }
-    return orderItemDTOS;
+    return orderItemResponseDTOList;
+  }
+
+  private void saveOrderItems(List<Map<String, String>> orderItems, Order order) {
+    for (Map<String, String> orderItemMap : orderItems) {
+      OrderItemDTO orderItemDTO = OrderUtil.getOrderItemDTO(orderItemMap);
+      OrderItem orderItem = OrderItemMapper.INSTANCE.toOrderItem(orderItemDTO);
+      orderItem.setOrderId(order.getOrderId());
+      orderItem.setPrice(orderItem.getPrice());
+      orderItem.setProductId(orderItem.getProductId());
+      orderItem.setQuantity(orderItem.getQuantity());
+      orderItemRepository.save(orderItem);
+    }
   }
 }
